Produce the same Code2Vec path contexts on every run

+++ src/main/kotlin/astminer/common/model/PathContextsModel.kt
@@ -8,7 +8,7 @@ data class OrientedNodeType(val typeLabel: String, val direction: Direction)
 
 data class PathContext(val startToken: String, val orientedNodeTypes: List<OrientedNodeType>, val endToken: String)
 
-data class PathContextId(val startTokenId: Long, val pathId: Long, val endTokenId: Long)
+data class PathContextId(val startTokenId: Int, val pathId: Int, val endTokenId: Int)
 
 data class LabeledPathContexts<T>(val label: T, val pathContexts: Collection<PathContext>)
